+++ src/components/main/window/cloud/parts/Foto.tsx
@@ -4,7 +4,8 @@ import { ButtonSize } from '../../../../../types/components.type';
 import { deleteFileFromDatabase } from '../../../../../global/files-functions';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { deleteFile } from '../../../../../features/files-slice';
+import { deleteFoto } from '../../../../../features/foto-slice';
+import fileDownload from 'js-file-download';
 
 import Spinner from '../../../../common/spinner/Spinner';
 import Button from '../../../../common/button/Button';
@@ -24,19 +25,28 @@ const Foto = (props: Props) => {
     const [loading, setLoading] = useState<boolean>(false)
 
     const deleteHandle = async () => {
-        setLoading(true);
-        const income = await deleteFileFromDatabase(id);
-        setLoading(false);
+        if (window.confirm('Na pewno usunąć ?')) {
+            setLoading(true);
+            const income = await deleteFileFromDatabase(id);
+            setLoading(false);
 
-        if (income.status === true) {
-            dispatch(deleteFile(id))
-        } else {
-            navigate(`/error/${income.info}`)
+            if (income.status === true) {
+                dispatch(deleteFoto(id))
+            } else {
+                window.location.href = `/error/${income.info}`
+            }
         }
+
     }
 
     const downloadHandle = async () => {
-
+        try {
+            const res = await fetch(`http://localhost:8080/files/file/${id}`)
+            const blob = await res.blob()
+            fileDownload(blob, name)
+        } catch (err) {
+            window.location.href = `/error/${(err as Error).message}`
+        }     
     }
 
     const showPopupHandle = () => {
